Extracted all auth logic out of usersService
I noticed that there was a lot of auth logic in the usersService. The point when I realized that I needed to pull it out of usersService was when the BaseAPI needed access to some class that would be able to retrieve the refresh token so that we can replay requests.